Rename policy_testing package to policytest
